Make NPC list and NPC combatants reactive
This opened can of worms I didn't know about. We should refactor use of
entities so that entity is NOT data class. Also UI should use DTOs
instead of entities.
This will make it easier to make UI properly
reactive, because DTOs would be data classes, so checking whether data
needed by UI have changed is as simple as original == newValue (which is
default comparison Jetpack Compose uses for stuff like state). Also it
would avoid unnecessary UI redraws when unrelated property changes -
i.e. there is no reason to redraw list of NPCs when their wounds are
reduced, because that information is not visible in UI anyway.